First step
I wrote the first step of the protocol, i.e. the client send the first
message and the server salve it in its log. Also the client salve in
its log the message sent. Each state of the server and the client are
handled in two files with some function